Move tests to PlaceholdersUtilsTest
